Add '--lapliance-threshold' and '--save-invalid' options

+++ faceset_builder/face_collector/photo_collector.py
@@ -1,5 +1,4 @@
 import numpy as np
-import sys
 import os
 import cv2
 import face_recognition
@@ -11,8 +10,8 @@ from . import utils
 
 class Photo_Collector(Collector):
 
-    def __init__(self, target_faces, tolerance=0.5, min_face_size=256, crop_size=512, min_luminosity=10, max_luminosity=245, one_face=False, mask_faces=False):
-        Collector.__init__(self, target_faces, tolerance, min_face_size, crop_size, min_luminosity, max_luminosity, one_face, mask_faces )
+    def __init__(self, target_faces, tolerance=0.5, min_face_size=256, crop_size=512, min_luminosity=10, max_luminosity=245, laplacian_threshold=0, one_face=False, mask_faces=False, save_invalid=False):
+        Collector.__init__(self, target_faces, tolerance, min_face_size, crop_size, min_luminosity, max_luminosity, laplacian_threshold, one_face, mask_faces, save_invalid )
 
 
     def processPhotos(self, files, outdir, sample_height=500):
@@ -21,8 +20,6 @@ class Photo_Collector(Collector):
 
         for file in tqdm(files):
             counter += 1
-            #progress = (counter/total)*100;
-            #sys.stdout.write("\r{0:.3g}% \t".format(progress))
             
             img = cv2.imread(file)
             rgb_img = img[:, :, ::-1]
@@ -32,23 +29,30 @@ class Photo_Collector(Collector):
             self.processImage(img, sample, outfile)
 
 
-    def cleanDuplicates(self, files, outdir, tolerance):
+    def cleanDuplicates(self, files, tolerance):
         hashes = dict()
+
+        duplicate_dir = os.path.join(self._invalid_dir, "duplicates")
+        corrupted_dir = os.path.join(self._invalid_dir, "corrupted")
+
+        os.makedirs(duplicate_dir, exist_ok=True)
+        os.makedirs(corrupted_dir, exist_ok=True)
 
         total = len(files)
         counter = 0
 
         for file in tqdm(files):
             counter += 1
-            #progress = (counter/total)*100;
-            #sys.stdout.write("\r{0:.3g}% \t".format(progress))
             
             img = cv2.imread(file)
             try:
                 hs = imutils.dhash(img)
             except:
                 #remove corrupted file
-                os.remove(file)
+                if self.save_invalid:
+                    os.rename(file, os.path.join(corrupted_dir, os.path.basename(file)))
+                else:
+                    os.remove(file)
                 continue
             h, w = imutils.cv_size(img)
             size = int(round(h*w))
@@ -59,17 +63,21 @@ class Photo_Collector(Collector):
                 if utils.get_num_bits_different(value[0], hs) <= tolerance:
                     if value[1] < size:
                         del hashes[key]
-                        name = "duplicateof_{0}".format(os.path.basename(file))
-                        os.remove(file)
-                        #os.replace(key, os.path.join(outdir,name))
+                        name = "duplicateof_{0}_{1}".format(os.path.basename(file), os.path.basename(key))
+                        if self.save_invalid:
+                            os.rename(key, os.path.join(duplicate_dir, name))
+                        else:
+                            os.remove(file)
                         smallest = False
                     else:
                         smallest = True
-                        duplicate_name = "duplicateof_{0}".format(os.path.basename(key))
+                        duplicate_name = "duplicateof_{0}_{1}".format(os.path.basename(key), os.path.basename(file))
                         
             if smallest:
-                os.remove(file)
-                #os.replace(file, os.path.join(outdir,duplicate_name))
+                if self.save_invalid:
+                    os.rename(file, os.path.join(duplicate_dir, duplicate_name))
+                else:
+                    os.remove(file)
             else:
                 hashes[file] = [hs, size]
